refactor(derive): Generalize test name
